No move it to memorystore.py
