+++ python/ta/test/test_query.py
@@ -26,7 +26,7 @@ from typeagent.knowpro.interfaces import (
     Topic,
 )
 from typeagent.knowpro.kplib import KnowledgeResponse
-from typeagent.storage.memorystorage import MemoryStorageProvider
+from typeagent.storage.memorystore import MemoryStorageProvider
 from typeagent.knowpro.importing import ConversationSettings
 from typeagent.knowpro.query import (
     TextRangeSelector,
