+++ python/ta/test/test_storage_indexes.py
@@ -4,7 +4,7 @@
 import pytest
 
 from fixtures import needs_auth, storage, embedding_model  # type: ignore  # It's used!
-from typeagent.storage.memorystorage import MemoryStorageProvider
+from typeagent.storage.memorystore import MemoryStorageProvider
 
 
 @pytest.mark.asyncio
