+++ python/ta/typeagent/storage/memorystore.py
